--- parkingspot/src/main/java/parkingspot/services/SearchingService.java
@@ -1,16 +1,0 @@
-package parkingspot.services;
-
-
-
-import parkingspot.entities.Vehicle;
-import parkingspot.exception.ParkingException;
-
-public interface SearchingService {
-
-public void getRegNumberForColor(int level, String color) throws ParkingException;
-	
-	public void getSlotNumbersFromColor(int level, String colour) throws ParkingException;
-	
-	public int getSlotNoFromRegistrationNo(int level, String registrationNo) throws ParkingException;
-
-}
