+++ parkingspot/src/main/java/parkingspot/services/impl/ServicesImpl.java
@@ -20,7 +20,8 @@ import parkingspot.model.strategy.ParkingStrategy;
 import parkingspot.services.ParkingService;
 
 
-public class ParkingServiceImpl implements ParkingService
+
+public class ServicesImpl implements ParkingService
 {
 	private ParkingDataManager<Vehicle> dataManager = null;
 	
@@ -151,6 +152,81 @@ public class ParkingServiceImpl implements ParkingService
 		return value;
 	}
 	
+	@Override
+	public void getRegNumberForColor(int level, String color) throws ParkingException
+	{
+		lock.readLock().lock();
+		validateParkingLot();
+		try
+		{
+			List<String> registrationList = dataManager.getRegNumberForColor(level, color);
+			if (registrationList.size() == 0)
+				System.out.println("Not Found");
+			else
+				System.out.println(String.join(",", registrationList));
+		}
+		catch (Exception e)
+		{
+			throw new ParkingException(ErrorCode.PROCESSING_ERROR.getMessage(), e);
+		}
+		finally
+		{
+			lock.readLock().unlock();
+		}
+	}
+	
+
+
+	
+	@Override
+	public void getSlotNumbersFromColor(int level, String color) throws ParkingException
+	{
+		lock.readLock().lock();
+		validateParkingLot();
+		try
+		{
+			List<Integer> slotList = dataManager.getSlotNumbersFromColor(level, color);
+			if (slotList.size() == 0)
+				System.out.println("Not Found");
+			StringJoiner joiner = new StringJoiner(",");
+			for (Integer slot : slotList)
+			{
+				joiner.add(slot + "");
+			}
+			System.out.println(joiner.toString());
+		}
+		catch (Exception e)
+		{
+			throw new ParkingException(ErrorCode.PROCESSING_ERROR.getMessage(), e);
+		}
+		finally
+		{
+			lock.readLock().unlock();
+		}
+	}
+	
+	@Override
+	public int getSlotNoFromRegistrationNo(int level, String registrationNo) throws ParkingException
+	{ 
+		int value = -1;
+		lock.readLock().lock();
+		validateParkingLot();
+		try
+		{
+			value = dataManager.getSlotNoFromRegistrationNo(level, registrationNo);
+			System.out.println(value != -1 ? value : "Not Found");
+		}
+		catch (Exception e)
+		{
+			throw new ParkingException(ErrorCode.PROCESSING_ERROR.getMessage(), e);
+		}
+		finally
+		{
+			lock.readLock().unlock();
+		}
+		return value;
+	}
+
 	
 	@Override
 	public void doCleanup()
